requirements.txt
omdb==0.4.0
pytvdbapi==0.5.0
